==> jupyter_notebooks/1.2_Number_of_tRNAs_AGA_TTA.ipynb <==
import os
from Bio import SeqIO
import pandas as pd
from tqdm import tqdm

# Define the base directory
base_dir = '/input/all_genbank/'

# Prepare output data
records = []

# List all genus folders
genus_folders = [folder for folder in os.listdir(base_dir) if os.path.isdir(os.path.join(base_dir, folder))]

# Count total GenBank files for progress bar
total_files = sum(
    len([f for f in os.listdir(os.path.join(base_dir, genus)) if f.endswith(('.gbff', '.gbk'))])
    for genus in genus_folders
)

# Initialize progress bar
with tqdm(total=total_files, desc="Processing GenBank files") as pbar:
    for genus in genus_folders:
        genus_path = os.path.join(base_dir, genus)
        for file_name in os.listdir(genus_path):
            if file_name.endswith(('.gbff', '.gbk')):
                file_path = os.path.join(genus_path, file_name)
                gcf_id = "_".join(file_name.split("_")[:2])
                for record in SeqIO.parse(file_path, "genbank"):
                    organism = record.annotations.get("organism", "Unknown")
                    aga_count = 0
                    tta_count = 0
                    for feature in record.features:
                        if feature.type == "tRNA" and "anticodon" in feature.qualifiers:
                            anticodon = feature.qualifiers['anticodon'][0]
                            if "seq:tct" in anticodon.lower():  # AGA codon = anticodon 'TCT'
                                aga_count += 1
                            elif "seq:taa" in anticodon.lower():  # TTA codon = anticodon 'TAA'
                                tta_count += 1
                    records.append({
                        "Genus": genus,
                        "Organism": organism,
                        "GCF_ID": gcf_id,
                        "tRNA_AGA": aga_count,
                        "tRNA_TTA": tta_count
                    })
                pbar.update(1)

# Create DataFrame
trna_df = pd.DataFrame(records)

df_unique = trna_df.groupby(['Genus', 'Organism', 'GCF_ID'], as_index=False)[['tRNA_AGA', 'tRNA_TTA']].max()

df_unique.to_excel('/results/tRNA_per_genome.xlsx', index = False)

df_unique.head(50)
