==> jupyter_notebooks/4.2_TTA_genes_among_BGCs.ipynb <==
import pandas as pd

file_path =  "/input/antiSMASH_8/Actinokineospora/summary_antismash.csv"
antismash_df = pd.read_csv(file_path)

antismash_df.head()

columns_to_keep = ['Locus_ID', 'CDS_Product', 'Gene_Kind', 'Is_Core_Gene', 'Region_Nr',
       'Strain', 'Cand_Cluster_1', 'Cluster_Product_1']

antismash_df = antismash_df[columns_to_keep]

file_path =  "/results/Actinokineospora/all_first_TTA_Actinokineospora.csv"
tta_df = pd.read_csv(file_path)

tta_df.head()

# Reattempt the merge
merged_df = pd.merge(tta_df, antismash_df, on='Locus_ID', how='left')

merged_df.head(50)

file_path =  "/results/Actinokineospora/all_genes_TTA_Actinokineospora_bgc.csv"
merged_df.to_csv(file_path)

# Reattempt the merge
bgc_df = pd.merge(antismash_df, tta_df, on='Locus_ID', how='left')

bgc_df.columns

file_path =  "/results/Actinokineospora/UUA/antismash_TTA_Actinokineospora.csv"
bgc_df.to_csv(file_path)
