==> jupyter_notebooks/4.3_Analysis_BGC_vs_non_BGC.ipynb <==
import pandas as pd
import os

### Among TTA genes

file_path = '/results/Actinokineospora/all_genes_TTA_Actinokineospora_bgc.csv'
TTA_df = pd.read_csv(file_path, index_col = 'Locus_ID')

output_path = '/results/Actinokineospora/'

TTA_df = TTA_df.drop('Unnamed: 0', axis = 1)

TTA_df.head()

# Recalculate the proportions with raw numbers for BGC genes vs Non-BGC genes
proportion_bgc_df = TTA_df.groupby('GCF_ID').apply(
    lambda group: pd.Series({
        'Percentage_BGC_Genes': group['Gene_Kind'].notna().sum() / len(group) * 100,
        'BGC_Genes_Count': group['Gene_Kind'].notna().sum(),
        'Non_BGC_Genes_Count': group['Gene_Kind'].isna().sum(),
        'Organism': group['Organism'].iloc[0]  # Assuming the organism is the same for each GCF_ID
    })
).reset_index()

proportion_bgc_df

file_path = os.path.join(output_dir, 'BGC_among_TTA_percentage.csv')
proportion_bgc_df.to_csv(file_path)

### Among BGC genes

file_path = '/results/Actinokineospora/antismash_TTA_Actinokineospora.csv'
bgc_TTA_df = pd.read_csv(file_path, index_col = 'Locus_ID')

bgc_TTA_df

bgc_TTA_df = bgc_TTA_df.drop(['Unnamed: 0', 'Cand_Cluster_1'], axis = 1)

# Ensure both columns are strings
bgc_TTA_df["Region_Nr"] = bgc_TTA_df["Region_Nr"].astype(str)
bgc_TTA_df["Cluster_Product_1"] = bgc_TTA_df["Cluster_Product_1"].astype(str)

# Create the 'BGC' column
bgc_TTA_df["BGC"] = bgc_TTA_df["Region_Nr"] + "_" + bgc_TTA_df["Cluster_Product_1"]

organism_mapping = bgc_TTA_df.groupby("Strain")["Organism"].first()  # Get the first non-NaN Organism for each GCF_ID
bgc_TTA_df["Organism"] = bgc_TTA_df["Strain"].map(organism_mapping)  # Map the Organism back to the main dataframe

# Calculate the number of unique BGCs per strain
bgc_counts = bgc_TTA_df.groupby("Strain")["BGC"].nunique().reset_index(name="Total_BGCs")

# Calculate the number of BGCs with rare codons per strain
bgc_with_rare_codons = bgc_TTA_df[bgc_TTA_df["TTA Codon Count"] > 0].groupby("Strain")["BGC"].nunique().reset_index(name="BGCs_with_Rare_Codons")

# Calculate the total number of genes per strain
total_genes = bgc_TTA_df.groupby("Strain").size().reset_index(name="Total_Genes")

# Calculate the number of genes with rare codons per strain
genes_with_rare_codons = bgc_TTA_df[bgc_TTA_df["TTA Codon Count"] > 0].groupby("Strain").size().reset_index(name="Genes_with_Rare_Codons")

# Merge all metrics into a single DataFrame
result = bgc_counts.merge(bgc_with_rare_codons, on="Strain", how="left") \
                   .merge(total_genes, on="Strain", how="left") \
                   .merge(genes_with_rare_codons, on="Strain", how="left")

# Fill NaN values with 0 for strains with no rare codons
result.fillna(0, inplace=True)

# Convert counts to integers
result[["BGCs_with_Rare_Codons", "Genes_with_Rare_Codons"]] = result[["BGCs_with_Rare_Codons", "Genes_with_Rare_Codons"]].astype(int)

# Calculate percentages
result["Percentage_TTA_BGCs"] = (result["BGCs_with_Rare_Codons"] / result["Total_BGCs"]) * 100
result["Percentage_TTA_BGC_genes"] = (result["Genes_with_Rare_Codons"] / result["Total_Genes"]) * 100

# Add the Organism column directly from the original DataFrame
result["Organism"] = result["Strain"].map(bgc_TTA_df.set_index("Strain")["Organism"].to_dict())

# Rename column
result = result.rename(columns={"Strain": "GCF_ID"})

result = result.dropna()

result.head()

# Treat all organisms as the same genus
result["Genus"] = "Actinokineospora"

# Remove square brackets from strain names in the "Organism" column
result["Organism"] = result["Organism"].str.replace(r"[\[\]]", "", regex=True)

# Group organisms by the first two words in the "Organism" column
result["Strain"] = result["Organism"].apply(lambda x: " ".join(x.split()[:2]))

result.head()

file_path = os.path.join(output_dir, 'TTA_among_BGC_genes.csv')
result.to_csv(file_path)
