==> jupyter_notebooks/5.3_COG_TTA_containing_genes_position_0.ipynb <==
import pandas as pd
import os
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

file_path = '/input/eggnog/Actinokineospora/concatenated_annotations.csv'
concatenated_df = pd.read_csv(file_path)

concatenated_df.head()

# Load the dataframe 
file_path = '/results/Actinokineospora/cog_categories_summary_tta.csv'
cog_all_tta = pd.read_csv(file_path)

cog_all_tta.head()

# Load dataframe with TTA containing genes
file_path = '/results/Actinokineospora/all_first_TTA_Actinokineospora.csv'
TTA_df = pd.read_csv(file_path)

cog_df = pd.merge(UUA_df, concatenated_df, on = 'Locus_ID', how = 'left')

cog_df = cog_df[cog_df["Relative_Position"] <= 10]

def count_cog_categories_per_strain(df, cog_column="COG_category", strain_column="Strain"):
    """
    Count occurrences of each COG category for each strain.
    
    Parameters:
        df (pd.DataFrame): DataFrame containing the data.
        cog_column (str): Name of the column containing COG categories.
        strain_column (str): Name of the column containing strain names.
    
    Returns:
        pd.DataFrame: Summary DataFrame with counts of each COG category per strain.
    """
    if cog_column not in df.columns or strain_column not in df.columns:
        raise ValueError(f"'{cog_column}' or '{strain_column}' column not found in the DataFrame.")

    # Group by strain and COG category and count occurrences
    summary_df = (
        df.groupby([strain_column, cog_column])
        .size()
        .reset_index(name="Count")
        .pivot(index=strain_column, columns=cog_column, values="Count")
        .fillna(0)  # Fill missing values with 0
        .astype(int)  # Convert counts to integers
    )
    
    return summary_df

# Assuming the DataFrame is named `df`
cog_summary_tta_p0 = count_cog_categories_per_strain(cog_df, cog_column="COG_category", strain_column="Strain")

# Save the summary of number of genes in each COG category where TTA is in the first 10% within the gene length to a CSV file
output_file = '/results/Actinokineospora/cog_categories_summary_position_0.csv'
cog_summary_tta_p0.to_csv(output_file)
